pyproject.toml
[project]
name = "openproject-mcp"
version = "0.1.0"
description = ""
authors = [
    {name = "David",email = "davidtrufanov@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "dotenv (>=0.9.9,<0.10.0)",
    "requests (>=2.32.3,<3.0.0)",
    "mcp[cli] (>=1.9.3,<2.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-gigachat (>=0.3.10,<0.4.0)",
    "langchain-mcp-adapters (>=0.1.7,<0.2.0)",
    "langgraph (>=0.4.8,<0.5.0)",
    "python-telegram-bot (>=22.1,<23.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
